refactor(container): mv Drawer to tool scope
