+++ src/containers/tool/Drawer/styles/content/default_content.js
