+++ src/containers/tool/Drawer/styles/mobile_closer.js
